SEL: working on moving all custom functions into a custom file

+++ public/src/modules/GenericDataParser/src/server/generic_data_parser.R
@@ -1375,7 +1375,9 @@ uploadRawDataOnly <- function(metaData, lsTransaction, subjectData, serverPath, 
   serverFileLocation <- moveFileToExperimentFolder(fileStartLocation, experiment, recordedBy, lsTransaction, configList$fileServiceType, configList$externalFileService)
   if(!is.null(reportFilePath) && reportFilePath != "") {
     batchNameList <- unique(subjectData$"Corporate Batch ID")
-    registerReportFile(reportFilePath, batchNameList, reportFileSummary, recordedBy, configList, experiment, lsTransaction, annotationType)
+    if (configList$useCustomReportRegistration == "true") {
+      registerReportFile(reportFilePath, batchNameList, reportFileSummary, recordedBy, configList, experiment, lsTransaction, annotationType)
+    }
   }
   
   # Analysis group
@@ -1711,7 +1713,9 @@ uploadData <- function(metaData,lsTransaction,calculatedResults,treatmentGroupDa
   serverFileLocation <- moveFileToExperimentFolder(fileStartLocation, experiment, recordedBy, lsTransaction, configList$fileServiceType, configList$externalFileService)
   if(!is.null(reportFilePath) && reportFilePath != "") {
     batchNameList <- unique(calculatedResults$"Corporate Batch ID")
-    registerReportFile(reportFilePath, batchNameList, reportFileSummary, recordedBy, configList, experiment, lsTransaction, annotationType)
+    if (configList$useCustomReportRegistration) {
+      registerReportFile(reportFilePath, batchNameList, reportFileSummary, recordedBy, configList, experiment, lsTransaction, annotationType)
+    }
   }
   
   # Each analysisGroupID creates an analysis group
@@ -1966,73 +1970,6 @@ uploadData <- function(metaData,lsTransaction,calculatedResults,treatmentGroupDa
   }
   return(NULL)
 }
-registerReportFile <- function(reportFilePath, batchNameList, reportFileSummary, recordedBy, configList, 
-                               experiment, lsTransaction, annotationType) {
-  # Registers a report as a batch annotation
-  
-  require(RCurl)
-  require(rjson)
-  
-  annotationList <- list(
-    dnsAnnotation = list(
-      name = basename(reportFilePath),
-      contentType = annotationType,
-      #description = "report file",
-      #dateExpired = "",
-      owningUrl = paste0(configList$serverPath, "experiments/codename/", experiment$codeName),
-      owningAttribute = "ACAS_experiment_annotation_id",
-      showInline = "false",
-      createdByLogin = recordedBy
-    ))
-  
-  annotationList$dnsAnnotation$annotationEntities <- lapply(batchNameList, function(batchCode) {
-    list(summary = reportFileSummary,
-         entity = list(
-           entityClass = "BATCH",
-           #entityURL = "",
-           entityCorpName = batchCode))
-    })
-  
-  tryCatch({response <- postForm(configList$reportRegistrationURL,
-                       FILE=fileUpload(filename = reportFilePath),
-                       PAYLOAD_TYPE="JSON",
-                       PAYLOAD=toJSON(annotationList))
-            response <- fromJSON(response)
-  }, error = function(e) {
-    stop("There was an error uploading the file for batch annotation")
-  })
-  
-  locationState <- experiment$lsStates[lapply(experiment$lsStates, function(x) x$"lsKind")=="report locations"]
-  
-  # Record the location
-  if (length(locationState)> 0) {
-    locationState <- locationState[[1]]
-  } else {
-    locationState <- createExperimentState(
-      recordedBy=recordedBy,
-      experiment = experiment,
-      lsType="metadata",
-      lsKind="report locations",
-      lsTransaction=lsTransaction)
-    
-    locationState <- saveExperimentState(locationState)
-  }
-  
-  tryCatch({
-    locationValue <- createStateValue(recordedBy = recordedBy,
-                                      lsType = "numericValue",
-                                      lsKind = "annotation id",
-                                      numericValue = response$dnsAnnotation$id,
-                                      lsState = locationState,
-                                      lsTransaction = lsTransaction)
-    
-    saveExperimentValues(list(locationValue))
-  }, error = function(e) {
-    stop("Could not save the annotation location")
-  })
-  
-  file.remove(reportFilePath)
-}
 runMain <- function(pathToGenericDataFormatExcelFile, reportFilePath=NULL, serverPath,
                     lsTranscationComments=NULL, dryRun, developmentMode = FALSE, testOutputLocation="./JSONoutput.json",
                     configList, testMode = FALSE, recordedBy) {
@@ -2180,8 +2117,11 @@ runMain <- function(pathToGenericDataFormatExcelFile, reportFilePath=NULL, serve
   
   # Delete any old data under the same experiment name (delete and reload)
   if(!dryRun && !newExperiment && errorFree) {
-    deleteSourceFile(experiment, configList)
-    deleteAnnotation(experiment, configList)
+    if(configList$deleteOnReload == "true") {
+      deleteSourceFile(experiment, configList)
+      deleteAnnotation(experiment, configList)
+    }
+
     deletedExperimentCodes <- c(experiment$codeName, getPreviousExperimentCodes(experiment))
     deleteExperiment(experiment)
   } else {
@@ -2254,62 +2194,6 @@ runMain <- function(pathToGenericDataFormatExcelFile, reportFilePath=NULL, serve
   summaryInfo$experimentEntity <- experiment
   
   return(summaryInfo)
-}
-deleteAnnotation <- function(experiment, configList) {
-  locationState <- experiment$lsStates[lapply(experiment$lsStates, function(x) x$"lsKind")=="report locations"]
-  
-  # Record the location
-  if (length(locationState)> 0) {
-    locationState <- locationState[[1]]
-    
-    lsKinds <- lapply(locationState$lsValues, function(x) x$"lsKind")
-    
-    valuesToDelete <- locationState$lsValues[lsKinds %in% c("annotation id")]
-    
-    if (length(valuesToDelete) > 0) {
-      tryCatch({
-        response <- getURL(
-          paste0(configList$reportRegistrationURL, "delete/", valuesToDelete[[1]]$numericValue),
-          customrequest='DELETE',
-          httpheader=c('Content-Type'='application/json'),
-          postfields=toJSON(experiment))
-      }, error = function(e) {
-        stop("There was an error deleting the old experiment annotation. Please contact your system adminstrator.")
-      })
-      if(!grepl("Deleted Annotation", response)) {
-        stop (paste("The loader was unable to delete the old experiment annotation. Instead, it got this response:", response))
-      }
-    }
-  }
-}
-deleteSourceFile <- function(experiment, configList) {
-
-  require(RCurl)
-  require(rjson)
-  
-  locationState <- experiment$lsStates[lapply(experiment$lsStates, function(x) x$"lsKind")=="raw results locations"]
-  if (length(locationState) > 0) {
-    locationState <- locationState[[1]]
-    
-    lsKinds <- lapply(locationState$lsValues, function(x) x$"lsKind")
-    
-    valuesToDelete <- locationState$lsValues[lsKinds %in% c("source file")]
-    
-    if (length(valuesToDelete) > 0) {
-      fileToDelete <- valuesToDelete[[1]]$fileValue
-      tryCatch({
-        response <- getURL(
-          paste0(configList$externalFileService, "deactivate/", fileToDelete),
-                 customrequest='DELETE',
-                 httpheader=c('Content-Type'='application/json'))
-      }, error = function(e) {
-        stop("There was an error deleting the old source file. Please contact your system adminstrator.")
-      })
-      if(!grepl("^Deactivated DNSFile", response)) {
-        warning(paste("The loader was unable to delete the old experiment source file. Instead, it got this response:", response))
-      }
-    }
-  }
 }
 moveFileToExperimentFolder <- function(fileStartLocation, experiment, recordedBy, lsTransaction, fileServiceType, fileService) {
   # Creates a folder for the excel file that was parsed and puts the file there. Returns the new location.
